+++ CHANGES.rst
@@ -483,6 +483,9 @@ Bug Fixes
   - ``SkyCoord.to_string`` no longer gives an error when used with an array
     coordinate with more than one dimension. [#3340]
 
+  - Fixed support for subclasses of ``UnitSphericalRepresentation`` and
+    ``SphericalRepresentation`` [#3354, #3366]
+
 - ``astropy.cosmology``
 
 - ``astropy.io.ascii``